RavenDB-13110 - fixing PR comments #3
